+++ ChoreBuddies.Backend.sln
@@ -16,6 +16,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Editor Configuration", "Edi
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "ChoreBuddies_SharedModels", "libs\shared-models\ChoreBuddies_SharedModels\ChoreBuddies_SharedModels.shproj", "{851D259E-BA39-4E5F-894F-171E4DCA97F2}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -44,5 +46,9 @@ Global
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {79F706B6-7590-441F-BC71-08A8A8AE88D2}
+	EndGlobalSection
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		libs\shared-models\ChoreBuddies_SharedModels\ChoreBuddies_SharedModels.projitems*{149cb185-0fc2-41ea-f149-8dbc202b8d71}*SharedItemsImports = 5
+		libs\shared-models\ChoreBuddies_SharedModels\ChoreBuddies_SharedModels.projitems*{851d259e-ba39-4e5f-894f-171e4dca97f2}*SharedItemsImports = 13
 	EndGlobalSection
 EndGlobal
