+++ ChoreBuddies.Backend/ChoreBuddies.Backend.csproj
@@ -42,4 +42,6 @@
     <ProjectReference Include="..\ChoreBuddies.Backend.ServiceDefaults\ChoreBuddies.Backend.ServiceDefaults.csproj" />
   </ItemGroup>
 
+  <Import Project="..\libs\shared-models\ChoreBuddies_SharedModels\ChoreBuddies_SharedModels.projitems" Label="Shared" />
+
 </Project>
