+++ libs/shared-models/ChoreBuddies_SharedModels/ChoreBuddies_SharedModels.projitems
@@ -9,6 +9,9 @@
     <Import_RootNamespace>ChoreBuddies_SharedModels</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Folder Include="$(MSBuildThisFileDirectory)Chores\" />
+    <Compile Include="$(MSBuildThisFileDirectory)Chores\ChoreDto.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Chores\ChoreOverviewDto.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Chores\Status.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)DefaultChores\DefaultChoreDto.cs" />
   </ItemGroup>
 </Project>
