+++ libs/shared-models/ChoreBuddies_SharedModels/Chores/ChoreDto.cs
@@ -1,6 +1,4 @@
-﻿using ChoreBuddies.Backend.Features.Chores;
-
-namespace ChoreBuddies.Backend.Chores
+﻿namespace ChoreBuddies_SharedModels.Chores
 {
     public record ChoreDto(
     string Id,
